Move storage to client/lib
